chore: Day 21 framework hardening
1. Fix the pre-existing 1-line README.md formatting issue (missing blank
   line after the title) - a prerequisite for enabling format:check in CI.
2. Add npm run format:check as a CI step, and a conservative scheduled
   trigger (daily cron) that re-runs the same checks against the live
   Petstore API - no alerting/notification wiring added.
3. Add docs/API-BEHAVIOR-NOTES.md, centralizing the live-verified
   spec-vs-behavior findings from Days 1-20 (previously scattered across
   individual spec-file comments) into one indexed reference.
4. Extract src/data/safe-id.ts and reuse it in buildPet/buildOrder/buildUser,
   replacing 3x duplicated safe-integer-id logic.
5. Tighten the id field to z.number().int() in PetSchema, OrderSchema, and
   UserSchema, closing a previously-flagged (Day 4) contract-fidelity gap.

Follow-up fix: OrderSchema.id's tightening correctly rejects the
occasionally-oversized (> Number.MAX_SAFE_INTEGER) id the live server
assigns when POST /store/order is sent a minimal {} payload - a real
instance of the int64/float64 precision trap already documented for Pet.
Rather than weakening OrderSchema, tests/store/place-order.spec.ts's
minimal-payload test now validates the response via
OrderSchema.omit({ id: true }).parse(...), since that id's exact value was
never what the test was verifying.

+++ src/schemas/pet.schema.ts
@@ -3,7 +3,7 @@ import { z } from 'zod';
 export const PetStatusSchema = z.enum(['available', 'pending', 'sold']);
 
 export const PetSchema = z.object({
-  id: z.number().optional(),
+  id: z.number().int().optional(),
   name: z.string(),
   category: z
     .object({
